# Speed up YAML parsing with CSafeLoader