yaml.FullLoader is quite slow and not safe by default (can execute code), use CSafeLoader (which uses a C parser, CLoader is equivalent in parsing time)

Time to create docs before: 14 seconds
TIme to create docs after: 2 seconds

Loading a single torque tune takes ~80 ms before this change (on PC), now it's ~10 ms