@adeebshihadeh 2022-06-24
Looks good, but a better title would be something like "Speed up YAML parsing with CSafeLoader"