@adeebshihadeh 2022-06-24
We can also just as easily make this python code instead of yaml.