@sshane 2022-06-24
I think it might be easier to export yaml than it is to automatically build valid Python, but up to @HaraldSchafer